+++ packages/functions/src/index.ts
@@ -1,3 +1,1 @@
-export * from './awsCredentialsProvider';
-export * from './getVercelOidcToken';
 export * from './waitUntil';
