+++ packages/functions/src/oidc/index.ts
@@ -1,0 +1,2 @@
+export * from './awsCredentialsProvider';
+export * from './getVercelOidcToken';
