+++ packages/functions/test/oidc/getVercelOidcToken.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it, beforeEach, afterEach } from 'vitest';
-import { getVercelOidcToken } from '../src';
+import { getVercelOidcToken } from '../../src/oidc';
 import { randomUUID } from 'crypto';
 
 describe('when VERCEL_OIDC_TOKEN is present in the environment variables', () => {
@@ -12,8 +12,8 @@ describe('when VERCEL_OIDC_TOKEN is present in the environment variables', () =>
     delete process.env.VERCEL_OIDC_TOKEN;
   });
 
-  it('returns the OIDC token', () => {
-    expect(getVercelOidcToken()).toEqual(token);
+  it('returns the OIDC token', async () => {
+    await expect(getVercelOidcToken()).resolves.toEqual(token);
   });
 });
 
@@ -39,14 +39,14 @@ describe('when loading from the request context', () => {
     ];
   });
 
-  it('returns the OIDC token', () => {
-    expect(getVercelOidcToken()).toEqual(token);
+  it('returns the OIDC token', async () => {
+    await expect(getVercelOidcToken()).resolves.toEqual(token);
   });
 });
 
 describe('when neither the environment variables or the request context is present', () => {
-  it('throws an error', () => {
-    expect(() => getVercelOidcToken()).toThrow(
+  it('throws an error', async () => {
+    await expect(getVercelOidcToken()).rejects.toThrow(
       /The 'x-vercel-oidc-token' header is missing from the request/
     );
   });
